Split listen_socket into its own files
In working on testing the server sockets for issue #5, the basesock
file needed a whole bunch of dependencies for the listen_socket which
were unreasonable.  So, split the listen_socket out into its own
files, for simplicity.

Also found a small build problem, in which the console module was
being compiled as a module, but was being linked into the r9d at
link time.  It's now a module, and we'll have to add some code
to load the lib and all that only if consoles are defined in the
config.  The setup_console and cleanup_console routines in server.cc
are currently stubbed out, so at least we're not breaking any
existing code.
