+++ server/classes/basesock.h
@@ -1,6 +1,6 @@
 /* basesock.h                                              -*- C++ -*-
  *   by Trinity Quirk <tquirk@ymb.net>
- *   last updated 24 Jul 2015, 13:03:39 tquirk
+ *   last updated 01 Nov 2015, 10:26:50 tquirk
  *
  * Revision IX game server
  * Copyright (C) 2015  Trinity Annabelle Quirk
@@ -20,7 +20,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
  *
- * This file contains a basic listening socket, with a listener thread.
+ * This file contains a base socket type.
  *
  * Things to do
  *
@@ -30,14 +30,11 @@
 #define __INC_BASESOCK_H__
 
 #include <netinet/in.h>
-#include <pthread.h>
+#include <netdb.h>
 
 #include <cstdint>
-#include <map>
 
 #include "control.h"
-#include "thread_pool.h"
-#include "defs.h"
 
 class basesock
 {
@@ -81,38 +78,6 @@ class base_user {
     virtual bool operator==(const base_user&) const;
 
     virtual const base_user& operator=(const base_user&);
-};
-
-class listen_socket {
-  public:
-    static const int REAP_TIMEOUT = 15;
-    static const int PING_TIMEOUT = 30;
-    static const int LINK_DEAD_TIMEOUT = 75;
-
-  protected:
-    pthread_t reaper;
-
-  public:
-    std::map<uint64_t, base_user *> users;
-    ThreadPool<packet_list> *send_pool;
-    ThreadPool<access_list> *access_pool;
-    basesock sock;
-
-  public:
-    listen_socket(struct addrinfo *, uint16_t);
-    virtual ~listen_socket();
-
-    void init(void);
-
-    virtual void start(void) = 0;
-    virtual void stop(void);
-
-    static void *access_pool_worker(void *);
-
-    virtual void login_user(access_list&);
-    virtual void logout_user(access_list&);
-
-    virtual void do_login(uint64_t, Control *, access_list&) = 0;
 };
 
 #endif /* __INC_BASESOCK_H__ */
