+++ server/classes/dgram.h
@@ -1,6 +1,6 @@
 /* dgram.h                                                 -*- C++ -*-
  *   by Trinity Quirk <tquirk@ymb.net>
- *   last updated 24 Jul 2015, 13:08:03 tquirk
+ *   last updated 01 Nov 2015, 11:24:46 tquirk
  *
  * Revision IX game server
  * Copyright (C) 2015  Trinity Annabelle Quirk
@@ -35,7 +35,7 @@
 #include <map>
 
 #include "control.h"
-#include "basesock.h"
+#include "listensock.h"
 #include "sockaddr.h"
 
 class dgram_user : public base_user
